Move replay controls to ReplayOverlay
