+++ packages/vscode-extension/src/webview/components/ReplayOverlay.css
@@ -4,7 +4,7 @@
   left: 0;
   width: 100%;
   height: 100%;
-  background-color: rgba(0, 0, 0, 0.7);
+  background: linear-gradient(0deg, rgba(0, 0, 0, 1), rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5),rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5));
   display: flex;
   justify-content: center;
   align-items: center;
@@ -68,7 +68,6 @@
 }
 
 .replay-close {
-  display: flex;
   position: absolute;
   top: 30px;
   right: 30px;
@@ -80,4 +79,105 @@
   position: absolute;
   bottom: 30px;
   right: 30px;
+}
+
+.replay-controls {
+  position: absolute;
+  width: 70%;
+  bottom: 0;
+  display: flex;
+  padding-bottom: 2%;
+  flex-direction: row;
+  gap: 5px;
+  align-items: center;
+}
+
+.replay-controls-pad {
+  width: 0%;
+}
+
+.replay-controls .button:hover {
+  background-color: rgba(0, 0, 0, 0.2);
+}
+
+.len-select-trigger {
+  box-sizing: border-box;
+  display: inline-flex;
+  align-items: center;
+  padding: 0 11px;
+  font-size: 13px;
+  line-height: 1;
+  overflow: hidden;
+  cursor: pointer;
+  height: 36px;
+  border: 0px solid transparent;
+  border-radius: 18px;
+  gap: 5px;
+  color: var(--swm-replay-len-select);
+  background-color: var(--swm-replay-len-select-background);
+  user-select: none;
+  width: 60px;
+}
+
+.len-select-trigger:hover {
+  background-color: var(--swm-replay-len-select-hover-background);
+}
+
+.len-select-trigger[data-state="open"] {
+  border-radius: 18px 18px 18px 18px;
+  box-shadow: var(--swm-focus-outline);
+}
+
+.len-select-trigger:focus {
+  box-shadow: var(--swm-focus-outline);
+}
+
+.len-select-content {
+  overflow: hidden;
+  background-color: var(--swm-replay-len-select-background);
+  border-radius: 18px 18px 18px 18px;
+  transform: translateY(4px);
+  z-index: 10;
+}
+
+.len-select-viewport {
+  padding: 6px;
+}
+
+.len-select-item {
+  font-size: 13px;
+  line-height: 1.2;
+  display: flex;
+  align-items: center;
+  min-height: 30px;
+  height: auto;
+  padding: 0 4px;
+  position: relative;
+  border-radius: 4px;
+  color: var(--swm-default-text);
+}
+
+.len-select-item[data-highlighted] {
+  outline: none;
+  background-color: var(--swm-replay-len-select-highlighted-background);
+  cursor: pointer;
+}
+
+.seekbar {
+  position: relative;
+  display: flex;
+  flex-grow: 1;
+  height: 4px;
+  background-color: #aaa;
+  border-radius: 2px;
+  cursor: pointer;
+}
+
+.progress-bar {
+  position: absolute;
+  top: 0;
+  left: 0;
+  height: 100%;
+  background-color: #fff;
+  border-radius: 2px;
 }
