+++ packages/vscode-extension/src/webview/components/ReplayUI.css
@@ -1,151 +1,3 @@
-.replay-controls {
-  position: absolute;
-  background: linear-gradient(0deg, rgb(0% 0% 0%) 0%, rgb(0.391% 0.391% 0.391% / 0.99609375) 6.25%, rgb(1.563% 1.563% 1.563% / 0.984375) 12.5%, rgb(3.516% 3.516% 3.516% / 0.96484375) 18.75%, rgb(6.25% 6.25% 6.25% / 0.9375) 25%, rgb(9.766% 9.766% 9.766% / 0.90234375) 31.25%, rgb(14.063% 14.063% 14.063% / 0.859375) 37.5%, rgb(19.141% 19.141% 19.141% / 0.80859375) 43.75%, rgb(25% 25% 25% / 0.75) 50%, rgb(31.641% 31.641% 31.641% / 0.68359375) 56.25%, rgb(39.063% 39.063% 39.063% / 0.609375) 62.5%, rgb(47.266% 47.266% 47.266% / 0.52734375) 68.75%, rgb(56.25% 56.25% 56.25% / 0.4375) 75%, rgb(66.016% 66.016% 66.016% / 0.33984375) 81.25%, rgb(76.563% 76.563% 76.563% / 0.234375) 87.5%, rgb(87.891% 87.891% 87.891% / 0.12109375) 93.75%, rgb(100% 100% 100% / 0) 100% );
-  width: 100%;
-  bottom: 0;
-  display: flex;
-  padding-bottom: 5%;
-  padding-top: 15%;
-  flex-direction: row;
-	gap: 5px;
-  align-items: center;
-	/* opacity: 0; */
-	transition: opacity 0.3s ease;
-}
-
-.phone-screen:hover .replay-controls {
-  opacity: 1;
-}
-
-.replay-controls-pad {
-	width: 0%;
-}
-
-.replay-controls .button:hover {
-  background-color: rgba(0, 0, 0, 0.2);
-}
-
-.len-select-trigger {
-  box-sizing: border-box;
-  display: inline;
-  align-items: center;
-  padding: 0 10px;
-  font-size: 13px;
-  line-height: 1;
-  overflow: hidden;
-  cursor: pointer;
-  height: 36px;
-  border: 0px solid transparent;
-  border-radius: 18px;
-  gap: 5px;
-  color: var(--swm-len-select);
-  background-color: var(--swm-len-select-background);
-  user-select: none;
-  min-width: var(--len-select-min-width);
-}
-
-.len-select-trigger:hover {
-  background-color: var(--swm-len-select-hover-background);
-}
-
-.len-select-trigger[data-state="open"] {
-  border-radius: 18px 18px 18px 18px;
-  box-shadow: var(--swm-focus-outline);
-}
-
-.len-select-trigger[data-disabled] {
-  background-color: var(--swm-len-select-disabled-background);
-  color: var(--swm-secondary-text);
-  pointer-events: none;
-}
-
-.len-select-trigger:focus {
-  box-shadow: var(--swm-focus-outline);
-}
-
-.len-select-content {
-  overflow: hidden;
-  background-color: var(--swm-len-select-background);
-  border-radius: 18px 18px 18px 18px;
-  transform: translateY(4px);
-  padding-bottom: 4px;
-  max-width: var(--radix-select-content-available-width);
-  max-height: var(--radix-select-content-available-height);
-}
-
-.len-select-viewport {
-  padding: 6px;
-}
-
-.len-select-scroll {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.len-select-label {
-  padding: 0 4px 0 4px;
-  font-size: 13px;
-  line-height: 20px;
-}
-
-.len-select-separator {
-  height: 1px;
-  background-color: var(--swm-len-select-separator);
-  margin: 5px;
-}
-
-.len-select-item {
-  font-size: 13px;
-  line-height: 1.2;
-  display: flex;
-  align-items: center;
-  min-height: 30px;
-  height: auto;
-  padding: 0 4px;
-  position: relative;
-  border-radius: 4px;
-  color: var(--swm-default-text);
-}
-
-.len-select-item[data-highlighted] {
-  outline: none;
-  background-color: var(--swm-len-select-highlighted-background);
-  cursor: pointer;
-}
-
-.len-select-item-text {
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  max-width: calc(var(--radix-select-content-available-width) - 15px);
-}
-
-.len-select-item:hover .len-select-item-text {
-  white-space: normal;
-  padding-top: 6px;
-  padding-bottom: 6px;
-}
-
-.seekbar {
-  position: relative;
-  display: flex;
-  flex-grow: 1;
-  height: 4px;
-  background-color: #aaa;
-  border-radius: 2px;
-  cursor: pointer;
-}
-
-.progress-bar {
-  position: absolute;
-  top: 0;
-  left: 0;
-  height: 100%;
-  background-color: #fff;
-  border-radius: 2px;
-}
-
 .vhs-noise {
 	position: fixed;
 	top: 0;
