+++ packages/vscode-extension/src/webview/styles/theme.css
@@ -1,8 +1,10 @@
 :root {
   /* Background */
-  --background-dark-120: #1c1f2b;
-  --background-dark-100: #232736;
-  --background-dark-90: #272b3c;
+  --background-dark-160: #14171f;
+  --background-dark-140: #181b25;
+  --background-dark-120:#1c1f2b;
+  --background-dark-100: #222634;
+  --background-dark-90: #262a3b;
   --background-dark-80: #30354a;
   --background-dark-70: #383e58;
   --background-dark-60: #404867;
@@ -399,4 +401,10 @@ body[data-vscode-theme-kind="vscode-dark"] {
   --swm-zoom-select-highlighted: var(--background-dark-60);
   --swm-zoom-select-disabled-text: var(--background-dark-60);
   --swm-zoom-select-separator: var(--background-dark-50);
+
+  /* Replay len select */
+    --swm-replay-len-select: var(--off-white);
+    --swm-replay-len-select-background: var(--background-dark-160);
+    --swm-replay-len-select-hover-background: var(--background-dark-120);
+    --swm-replay-len-select-highlighted-background: var(--background-dark-80);
 }
